requirements.txt
flet[all]
pyserial

